update test, typo

+++ php/removeClassSection.php
@@ -7,6 +7,12 @@
 			printf("Connect failed: %s\n", mysqli_connect_error());
 			return "3";
 		}
+		$query = "SELECT * FROM `classSectionChange` WHERE `class`='$className' AND `username`='$userName' AND `secfrom`='$secFrom' AND `secto`='$secTo' LIMIT 1";
+		// DELETE UN-EXIST CLASS OR FROM UN-EXIST USER
+		if(mysqli_num_rows(mysqli_query($link, $query)) <= 0) {
+   			return "1";
+  		}
+
 		$query = "DELETE FROM `classSectionChange` WHERE `username`='$userName' AND `class`='$className' AND `secfrom`='$secFrom' AND `secto`='$secTo'";
 		if(!mysqli_query($link, $query)) {
 			return "3"; //DB ERROR
